P3 preserve json-c-derived array cardinality width
json_object_array_length() returns size_t, but the JSON walker, query-metrics warning path, WebSocket JSON-RPC batch path, and downstream ACLK capability lookup narrowed cardinalities or indexes to int. The defects are potential only above INT_MAX; normal health, MQTT, MCP, WebSocket, and ACLK environment payloads are far smaller.

Keep every affected length, count, bound, and index as size_t through comparison and indexing. The JSON callback stream, MCP warning branch, WebSocket batch dispatch, and ACLK protocol-capability lookup retain the same accepted inputs, normal iteration/order, return values, errors, output, ownership, locking, lifetime, and concurrency behavior. Native-width locals preserve O(n) work with zero meaningful performance cost.

Validated with full Debug and AddressSanitizer builds, focused JSON and full unit suites, changed-line clang-tidy, targeted cppcheck, exact same-failure scans, and six structured full-scope source-review approvals on the complete four-file diff.

(cherry picked from commit 4da2c9c50cdc638e580ffb118fbc0a96844ebb3a)

+++ src/libnetdata/json/json.c
@@ -188,10 +188,10 @@ static inline void json_jsonc_set_integer(JSON_ENTRY *e, char *key, int64_t valu
  * @param callback_function function used to create a silencer.
  */
 static inline void json_jsonc_parse_array(json_object *ptr, void *callback_data,int (*callback_function)(struct json_entry *)) {
-    int end = json_object_array_length(ptr);
+    size_t end = json_object_array_length(ptr);
 
     if(end) {
-        int i;
+        size_t i;
         i = 0;
 
         enum json_type type;
